split tangofile tests from tangoattr tests
