+++ .github/workflows/tests.yml
@@ -48,6 +48,8 @@ jobs:
         run: .ci/run.sh 3 diffractogram
       - name: run tangosource tests
         run: .ci/run.sh 3 tangosource
+      - name: run tangofilesource tests
+        run: .ci/run.sh 3 tangofilesource
       - name: run httpsource tests
         run: .ci/run.sh 3 httpsource
       - name: Stop the docker
